fix(CodeExpression): Fix hover state issues when reflowing with text
Hovering over a CodeExpression element near the edge of a container would cause the component to drop to the enxt line, in the process losing focus for a split second. This would happen repeatedly causing flickering. This change adds a time-delay before removing the hover state changes, giving a user time to interact with the component.