refactor(Editor): Simplify activeLanguage prop to make it unidirectional
Changes to the `activeLanguage` prop no longer emit the `setLanguage` custom event,
and don't mutate the prop variable.
Changing the language from the editor syntax picker dropdown emits the `setLanguage`
custom event as before, but it is up to the listening/parent component to update
any state and set the appropriate new language `prop` on the `editor` component.

Note that the `getContents` method will always use the latest value from the
language picker dropdown, and that the picker value will update whenever the prop
is changed on the `editor` component.